+++ app/db/models/request.py
@@ -15,7 +15,7 @@
 #
 
 
-from sqlalchemy import Column, BigInteger, Boolean, ForeignKey, Float, String
+from sqlalchemy import Column, BigInteger, Boolean, ForeignKey, Double, String
 from sqlalchemy.orm import relationship
 
 from app.db.base_class import Base
@@ -39,18 +39,18 @@ class Request(Base):
 
     input_method_id = Column(BigInteger, ForeignKey('methods.id', ondelete='SET NULL'), nullable=True)
     input_method = relationship('Method', foreign_keys=input_method_id, uselist=False, lazy='selectin')
-    input_value = Column(Float(), nullable=True)
+    input_value = Column(Double(), nullable=True)
 
-    value = Column(Float(), nullable=True)
-    input_rate = Column(Float(), nullable=True)
-    rate = Column(Float(), nullable=True)
-    output_rate = Column(Float(), nullable=True)
+    value = Column(Double(), nullable=True)
+    input_rate = Column(Double(), nullable=True)
+    rate = Column(Double(), nullable=True)
+    output_rate = Column(Double(), nullable=True)
 
     output_method_id = Column(BigInteger, ForeignKey('methods.id', ondelete='SET NULL'), nullable=True)
     output_method = relationship('Method', foreign_keys=output_method_id, uselist=False, lazy='selectin')
     output_requisite_data_id = Column(BigInteger, ForeignKey('requisites_datas.id', ondelete='SET NULL'), nullable=True)
     output_requisite_data = relationship('RequisiteData', foreign_keys=output_requisite_data_id,
                                          uselist=False, lazy='selectin')
-    output_value = Column(Float(), nullable=True)
+    output_value = Column(Double(), nullable=True)
 
     is_deleted = Column(Boolean, default=False)
