+++ app/db/models/transfer.py
@@ -15,7 +15,7 @@
 #
 
 
-from sqlalchemy import Column, Boolean, ForeignKey, BigInteger, Float
+from sqlalchemy import Column, Boolean, ForeignKey, BigInteger, Double
 from sqlalchemy.orm import relationship
 
 from app.db.base_class import Base
@@ -30,5 +30,5 @@ class Transfer(Base):
     wallet_from = relationship('Wallet', foreign_keys=wallet_from_id, uselist=False, lazy='selectin')
     wallet_to_id = Column(BigInteger, ForeignKey('wallets.id', ondelete='SET NULL'), nullable=True)
     wallet_to = relationship('Wallet', foreign_keys=wallet_to_id, uselist=False, lazy='selectin')
-    value = Column(Float(), default=0)
+    value = Column(Double(), default=0)
     is_deleted = Column(Boolean, default=False)
