+++ app/db/models/wallet.py
@@ -15,7 +15,7 @@
 #
 
 
-from sqlalchemy import Column, BigInteger, Boolean, String, Float
+from sqlalchemy import Column, BigInteger, Boolean, String, Double
 
 from app.db.base_class import Base
 
@@ -26,7 +26,7 @@ class Wallet(Base):
     id = Column(BigInteger, primary_key=True)
 
     name = Column(String(32), default='Wallet')
-    value = Column(Float(), default=0)
-    value_banned = Column(Float(), default=0)
-    value_can_minus = Column(Float(), default=0)
+    value = Column(Double(), default=0)
+    value_banned = Column(Double(), default=0)
+    value_can_minus = Column(Double(), default=0)
     is_deleted = Column(Boolean, default=False)
